Fix ObjectComparison
